backend/requirements.txt
# Django Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Data Processing
pandas>=2.1,<3.0
numpy>=1.26,<2.0
openpyxl>=3.1,<4.0

# Utilities
python-dateutil>=2.8,<3.0
python-decouple>=3.8,<4.0

# Production Server
gunicorn>=21.0,<22.0
